Issue 3811: Fix SpotBugs error on Java 11 build (#3812)
* Fixes BX_UNBOXING_IMMEDIATELY_REBOXED spotBugs error observed on Java 11 based build.

* Removes Javadoc warnings in Pravega client build.

+++ client/src/main/java/io/pravega/client/stream/impl/ReaderGroupImpl.java
@@ -222,7 +222,7 @@ public class ReaderGroupImpl implements ReaderGroup, ReaderGroupMetrics {
                          .flatMap(map -> map.entrySet().stream())
                          .collect(Collectors.toMap(Entry::getKey,
                                                    // A value of -1L implies read until the end of the segment.
-                                                   entry -> (entry.getValue() == -1L) ? Long.MAX_VALUE : entry.getValue()));
+                                                   entry -> (entry.getValue() == -1L) ? (Long) Long.MAX_VALUE : entry.getValue()));
     }
 
     @Override
